# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion de biens immobilier",
        "description": "Documentation de l'API immoApp",
        "version": "1.0.0"
    },
    "paths": {
...
        "/type-payements": {
            "get": {
                "tags": [
                    "TypePayements"
                ],
                "summary": "Liste tous les types de payement",
                "operationId": "4d8f334b0c4b8177b9cf9320f0d887ad",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TypePayements"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "TypePayements"
                ],
                "summary": "Créer un type de payement",
                "operationId": "d48cf303c11551597f308ff0e59cdd7f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TypePayements"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Type créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TypePayements"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/type-payements/{id}": {
            "get": {
                "tags": [
                    "TypePayements"
                ],
                "summary": "Afficher un type de payement",
                "operationId": "ead4ed3f7201a6bf79a31b837e91531d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Type trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TypePayements"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "TypePayements"
                ],
                "summary": "Mettre à jour un type de payement",
                "operationId": "3a41f7796c54bcccb6062ffd2f32ba81",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TypePayements"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Type mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TypePayements"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "TypePayements"
                ],
                "summary": "Supprimer un type de payement",
                "operationId": "3dabf900c5f28cfc35f5c4b33241f835",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Type supprimé"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
        "/type-utilisateurs": {
            "get": {
                "tags": [
                    "TypeUtilisateurs"
                ],
                "summary": "Liste tous les types d'utilisateur",
                "operationId": "86b66fad2031b9becda867cd51025a37",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TypeUtilisateurs"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "TypeUtilisateurs"
                ],
                "summary": "Créer un type d'utilisateur",
                "operationId": "df7f42a46eeffa125da36d6367be3c1f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TypeUtilisateurs"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Type créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TypeUtilisateurs"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/type-utilisateurs/{id}": {
            "get": {
                "tags": [
                    "TypeUtilisateurs"
                ],
                "summary": "Afficher un type d'utilisateur",
                "operationId": "134a8efb4d94a73054df8f3e3e8922bc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Type trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TypeUtilisateurs"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "TypeUtilisateurs"
                ],
                "summary": "Mettre à jour un type d'utilisateur",
                "operationId": "32fb6d71b1f60793a2ff5fdc22fdd1f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TypeUtilisateurs"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Type mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TypeUtilisateurs"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "TypeUtilisateurs"
                ],
                "summary": "Supprimer un type d'utilisateur",
                "operationId": "48be682f582acc501cb64004461819b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Type supprimé"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TypePayements": {
                "required": [
                    "libelle"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "libelle": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "TypeUtilisateurs": {
                "required": [
                    "libelle"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "libelle": {
                        "type": "string",
                        "example": "Propriétaire"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}